Work around compile error in CanvasUITests
refs: none
affects: none
release note: none
